av1_dist_block: quiet -Wshorten-64-to-32 warning
make tmp an int64 as 12-bit may exceed 32-bits

BUG=aomedia:445

+++ av1/encoder/rdopt.c
@@ -1424,7 +1424,7 @@ void av1_dist_block(const AV1_COMP *cpi, MACROBLOCK *x, int plane,
     const tran_low_t *dqcoeff = BLOCK_OFFSET(pd->dqcoeff, block);
     const uint16_t eob = p->eobs[block];
 
-    unsigned int tmp;
+    int64_t tmp;
 
     assert(cpi != NULL);
     assert(tx_size_wide_log2[0] == tx_size_high_log2[0]);
@@ -1455,14 +1455,15 @@ void av1_dist_block(const AV1_COMP *cpi, MACROBLOCK *x, int plane,
       const int diff_idx = (blk_row * diff_stride + blk_col)
                            << tx_size_wide_log2[0];
       const int16_t *diff = &p->src_diff[diff_idx];
-      tmp = sum_squares_visible(xd, plane, diff, diff_stride, blk_row, blk_col,
-                                plane_bsize, tx_bsize);
+      tmp = sum_squares_visible(xd, plane, diff, diff_stride,
+                                              blk_row, blk_col, plane_bsize,
+                                              tx_bsize);
 #if CONFIG_HIGHBITDEPTH
       if (xd->cur_buf->flags & YV12_FLAG_HIGHBITDEPTH)
         tmp = ROUND_POWER_OF_TWO(tmp, (xd->bd - 8) * 2);
 #endif  // CONFIG_HIGHBITDEPTH
     }
-    *out_sse = (int64_t)tmp * 16;
+    *out_sse = tmp * 16;
 
     if (eob) {
       if (output_status == OUTPUT_HAS_DECODED_PIXELS) {
